biginteger

+++ app/bank/models.py
@@ -23,9 +23,9 @@ class Agency(models.Model):
     
     bank_id = models.ForeignKey('Bank',on_delete=models.CASCADE)
     address = models.CharField(max_length=250,null=False)
-    phone = models.DecimalField(max_digits=11,decimal_places=0,unique=True,null=False)
+    phone = models.BigIntegerField(unique=True,null=False)
     phonetype = models.IntegerField(choices=TYPE_PHONE,null=False)
-    phone1 = models.DecimalField(max_digits=11,decimal_places=0,blank=True)
+    phone1 = models.BigIntegerField(blank=True)
     phonetype1 = models.IntegerField(choices=TYPE_PHONE,blank=True)
     agency = models.CharField(max_length=250,null=False)
     agency_name = models.CharField(max_length=250,null=False)
